Better types & typecasting in throttle control arithmetic.

+++ cl_libevent2/src/cl_cluster.c
@@ -1471,14 +1471,15 @@ node_throttle_control(cl_cluster_node* cn)
 	// Calculate the failure percentage. Work in units of tenths-of-a-percent
 	// for finer resolution of the resulting throttle percent.
 	uint64_t sum = failures_sum + successes_sum;
-	uint64_t failure_tenths_pct = sum == 0 ? 0 : (failures_sum * 1000) / sum;
+	uint32_t failure_tenths_pct =
+			sum == 0 ? 0 : (uint32_t)((failures_sum * 1000) / sum);
 
 	// TODO - anything special for a 100% failure rate? Several seconds of all
 	// failures with 0 successes might mean we should destroy this node?
 
 	// Calculate and apply the throttle rate.
 	uint32_t throttle_pct = 0;
-	uint64_t threshold_tenths_pct = threshold_failure_pct * 10;
+	uint32_t threshold_tenths_pct = threshold_failure_pct * 10;
 
 	if (enough_history && failure_tenths_pct > threshold_tenths_pct) {
 		throttle_pct = ((failure_tenths_pct - threshold_tenths_pct) *
@@ -1489,7 +1490,7 @@ node_throttle_control(cl_cluster_node* cn)
 		}
 	}
 
-	cf_debug("node %s recent successes %u, failures %u, failure-tenths-pct %u, throttle-pct %u",
+	cf_debug("node %s recent successes %lu, failures %lu, failure-tenths-pct %u, throttle-pct %u",
 			cn->name, successes_sum, failures_sum, failure_tenths_pct, throttle_pct);
 
 	cf_atomic32_set(&cn->throttle_pct, throttle_pct);
